App/Service: Replace gServiceId with DataStore's increamentalCounter
In the MainService, a created service requires a unique identifier. To
this end, gServiceId, a private class integer member, has been used.
However, a class member would be invalidated or reinitialized after the
instance is destroyed so a more durable solution to generate an
identifier number is needed· For this reason, this patch replaces
gServiceId with DataStore's increamentalCounter.

Signed-off-by: Wook Song <wook16.song@samsung.com>